gcn_stream first test
