+++ .gitignore
@@ -131,3 +131,4 @@ dmypy.json
 
 # Custom
 conf/local_conf.conf
+gcn_storage_test
