+++ fink_grb/online/gcn_stream.py
@@ -1,16 +1,18 @@
-import pandas as pd
 import argparse
 
 import pyarrow as pa
 import pyarrow.parquet as pq
 
+import voeventparse as vp
+
 from gcn_kafka import Consumer
 from instruments import FERMI, SWIFT, INTEGRAL, ICECUBE, LISTEN_PACKS, INSTR_SUBSCRIBES, detect_instruments
 
-import voeventparse as vp
 import io
 import os
 import configparser
+
+import gcn_reader as gr
 
 
 def getargs(parser: argparse.ArgumentParser) -> argparse.Namespace:
@@ -45,82 +47,6 @@ def getargs(parser: argparse.ArgumentParser) -> argparse.Namespace:
     return args
 
 
-def get_trigger_id(voevent):
-    toplevel_params = vp.get_toplevel_params(voevent)
-
-    if "TrigID" in toplevel_params:
-        return int(toplevel_params["TrigID"]["value"])
-
-    if "AMON_ID" in toplevel_params:
-        return int(toplevel_params["AMON_ID"]["value"])
-
-    return -1
-
-
-def voevent_to_df(voevent):
-
-
-    ivorn = voevent.attrib['ivorn']
-    instruments = detect_instruments(ivorn)
-
-    trigger_id = get_trigger_id(voevent)
-    
-    coords = vp.get_event_position(voevent)
-    time_utc = vp.get_event_time_as_utc(voevent)
-
-    if instruments == "Fermi":
-        error_unit  = "deg"
-    elif instruments == "SWIFT":
-        error_unit = "arcmin"
-    elif instruments == "INTEGRAL":
-        error_unit = "arcmin"
-    elif instruments == "ICECUBE":
-        error_unit = "deg"
-    else:
-        raise ValueError("bad instruments: {}".format(instruments))
-
-    df = pd.DataFrame.from_dict(
-        {
-            'instruments': [instruments],
-            'ivorn': [ivorn],
-            'trigger_id': [trigger_id],
-            'ra': [coords.ra],
-            'dec': [coords.dec],
-            'err': [coords.err],
-            'units': [error_unit],
-            'timeUTC': [time_utc],
-            'raw_event': vp.prettystr(voevent)
-        }
-    )
-
-    return df
-
-
-def is_observation(voevent):
-    gcn_role = voevent.attrib["role"]
-    return gcn_role == "observation"
-
-
-def is_listened_packets_types(voevent, listen_packs):
-    toplevel_params = vp.get_toplevel_params(voevent)
-    gcn_packet_type = toplevel_params["Packet_Type"]["value"]
-
-    if int(gcn_packet_type) in listen_packs:
-        return True
-
-    return False
-
-
-def load_voevent(file, verbose=False):
-    try:
-            v = vp.load(file)
-            return v
-    except Exception as e:
-        if verbose:
-            print("failed to load the voevent:\n\tlocation={}\n\tcause={}".format(file, e))
-        raise e
-
-
 if __name__ == "__main__":
 
     parser = argparse.ArgumentParser(description=__doc__)
@@ -145,16 +71,24 @@ if __name__ == "__main__":
             value = message.value()
             
             decode = io.BytesIO(value).read().decode("UTF-8")
-            voevent = load_voevent(decode)
+
+            try:
+                voevent = gr.load_voevent(io.StringIO(decode))
+            except Exception as e:
+                print(decode)
+                print(e)
+                print()
+                print()
+                continue
 
             toplevel_params = vp.get_toplevel_params(voevent)
             gcn_packet_type = toplevel_params["Packet_Type"]["value"]
 
             gcn_how_description = voevent.How.Description
 
-            if is_observation(voevent) and is_listened_packets_types(voevent, LISTEN_PACKS):
+            if gr.is_observation(voevent) and gr.is_listened_packets_types(voevent, LISTEN_PACKS):
 
-                df = voevent_to_df(voevent)
+                df = gr.voevent_to_df(voevent)
 
                 df['year'] = df['timeUTC'].dt.strftime('%Y')
                 df['month'] = df['timeUTC'].dt.strftime('%m')
